requirements.txt
numpy>=1.24
pandas>=1.5
matplotlib>=3.7
pyyaml>=6.0
gymnasium>=0.29
torch>=2.0
pytest>=7.0
